+++ StabilityMatrix.Avalonia/ViewModels/CheckpointBrowserViewModel.cs
@@ -1,14 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Net.Http;
+using System.Reactive;
+using System.Reactive.Linq;
+using System.Reactive.Threading.Tasks;
+using System.Threading;
 using System.Threading.Tasks;
 using AsyncAwaitBestPractices;
 using Avalonia.Collections;
 using Avalonia.Controls;
-using Avalonia.Controls.Notifications;
+using AvaloniaEdit.Utils;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
@@ -17,6 +22,7 @@ using Refit;
 using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
 using StabilityMatrix.Avalonia.ViewModels.CheckpointBrowser;
+using StabilityMatrix.Avalonia.ViewModels.CheckpointManager;
 using StabilityMatrix.Avalonia.Views;
 using StabilityMatrix.Core.Api;
 using StabilityMatrix.Core.Attributes;
@@ -28,6 +34,7 @@ using StabilityMatrix.Core.Models;
 using StabilityMatrix.Core.Models.Api;
 using StabilityMatrix.Core.Models.Settings;
 using StabilityMatrix.Core.Services;
+using Notification = Avalonia.Controls.Notifications.Notification;
 using Symbol = FluentIcons.Common.Symbol;
 using SymbolIconSource = FluentIcons.FluentAvalonia.SymbolIconSource;
 
@@ -60,10 +67,12 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
     [ObservableProperty] private bool hasSearched;
     [ObservableProperty] private bool canGoToNextPage;
     [ObservableProperty] private bool canGoToPreviousPage;
+    [ObservableProperty] private bool canGoToFirstPage;
+    [ObservableProperty] private bool canGoToLastPage;
     [ObservableProperty] private bool isIndeterminate;
     [ObservableProperty] private bool noResultsFound;
     [ObservableProperty] private string noResultsText = string.Empty;
-    [ObservableProperty] private string selectedBaseModelType = "All";    
+    [ObservableProperty] private string selectedBaseModelType = "All";
     
     private List<CheckpointBrowserCardViewModel> allModelCards = new();
     
@@ -94,6 +103,16 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
 
         CurrentPageNumber = 1;
         CanGoToNextPage = true;
+        CanGoToLastPage = true;
+
+        Observable
+            .FromEventPattern<PropertyChangedEventArgs>(this, nameof(PropertyChanged))
+            .Where(x => x.EventArgs.PropertyName == nameof(CurrentPageNumber))
+            .Throttle(TimeSpan.FromMilliseconds(250))
+            .Select(_ => CurrentPageNumber)
+            .Where(page => page <= TotalPages && page > 0)
+            .ObserveOn(SynchronizationContext.Current)
+            .Subscribe(_ => TrySearchAgain(false).SafeFireAndForget(), err => Logger.Error(err));
     }
 
     public override void OnLoaded()
@@ -101,6 +120,14 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
         if (Design.IsDesignMode) return;
         
         var searchOptions = settingsManager.Settings.ModelSearchOptions;
+        
+        // Fix SelectedModelType if someone had selected the obsolete "Model" option
+        if (searchOptions is {SelectedModelType: CivitModelType.Model})
+        {
+            settingsManager.Transaction(s => s.ModelSearchOptions = new ModelSearchOptions(
+                SelectedPeriod, SortMode, CivitModelType.Checkpoint, SelectedBaseModelType));
+            searchOptions = settingsManager.Settings.ModelSearchOptions;
+        }
         
         SelectedPeriod = searchOptions?.SelectedPeriod ?? CivitPeriod.Month;
         SortMode = searchOptions?.SortMode ?? CivitSortMode.HighestRated;
@@ -250,13 +277,21 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
                 }).ToList();
             
             allModelCards = updateCards;
-            ModelCards =
-                new ObservableCollection<CheckpointBrowserCardViewModel>(
-                    updateCards.Where(FilterModelCardsPredicate));
+
+            var filteredCards = updateCards.Where(FilterModelCardsPredicate);
+            if (SortMode == CivitSortMode.Installed)
+            {
+                filteredCards =
+                    filteredCards.OrderByDescending(x => x.UpdateCardText == "Update Available");
+            }
+            
+            ModelCards =new ObservableCollection<CheckpointBrowserCardViewModel>(filteredCards);
         }
         TotalPages = metadata?.TotalPages ?? 1;
+        CanGoToFirstPage = CurrentPageNumber != 1;
         CanGoToPreviousPage = CurrentPageNumber > 1;
         CanGoToNextPage = CurrentPageNumber < TotalPages;
+        CanGoToLastPage = CurrentPageNumber != TotalPages;
         // Status update
         ShowMainLoadingSpinner = false;
         IsIndeterminate = false;
@@ -309,6 +344,26 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
         {
             modelRequest.BaseModel = SelectedBaseModelType;
         }
+
+        if (SortMode == CivitSortMode.Installed)
+        {
+            var connectedModels =
+                CheckpointFile.GetAllCheckpointFiles(settingsManager.ModelsDirectory)
+                    .Where(c => c.IsConnectedModel);
+
+            if (SelectedModelType != CivitModelType.All)
+            {
+                connectedModels = connectedModels.Where(c => c.ModelType == SelectedModelType);
+            }
+
+            modelRequest = new CivitModelsRequest
+            {
+                CommaSeparatedModelIds = string.Join(",",
+                    connectedModels.Select(c => c.ConnectedModel!.ModelId).GroupBy(m => m)
+                        .Select(g => g.First())),
+                Types = SelectedModelType == CivitModelType.All ? null : new[] {SelectedModelType}
+            };
+        }
         
         // See if query is cached
         var cachedQuery = await liteDbContext.CivitModelQueryCache
@@ -344,22 +399,32 @@ public partial class CheckpointBrowserViewModel : PageViewModelBase
         UpdateResultsText();
     }
 
-    [RelayCommand]
-    private async Task PreviousPage()
+    public void FirstPage()
     {
-        if (CurrentPageNumber == 1) return;
-
-        CurrentPageNumber--;
-        await TrySearchAgain(false);
+        CurrentPageNumber = 1;
     }
 
-    [RelayCommand]
-    private async Task NextPage()
+    public void PreviousPage()
     {
-        CurrentPageNumber++;
-        await TrySearchAgain(false);
+        if (CurrentPageNumber == 1) 
+            return;
+        
+        CurrentPageNumber--;
     }
     
+    public void NextPage()
+    {
+        if (CurrentPageNumber == TotalPages) 
+            return;
+        
+        CurrentPageNumber++;
+    }
+
+    public void LastPage()
+    {
+        CurrentPageNumber = TotalPages;
+    }
+
     partial void OnShowNsfwChanged(bool value)
     {
         settingsManager.Transaction(s => s.ModelBrowserNsfwEnabled, value);