+++ StabilityMatrix.Avalonia/ViewModels/SettingsViewModel.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
+using System.Diagnostics;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -11,12 +13,14 @@ using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Controls.Notifications;
 using Avalonia.Styling;
+using Avalonia.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
 using NLog;
 using StabilityMatrix.Avalonia.Controls;
 using StabilityMatrix.Avalonia.Helpers;
+using StabilityMatrix.Avalonia.Languages;
 using StabilityMatrix.Avalonia.Models;
 using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
@@ -45,6 +49,7 @@ public partial class SettingsViewModel : PageViewModelBase
     private readonly IPrerequisiteHelper prerequisiteHelper;
     private readonly IPyRunner pyRunner;
     private readonly ServiceManager<ViewModelBase> dialogFactory;
+    private readonly ITrackedDownloadService trackedDownloadService;
     
     public SharedState SharedState { get; }
     
@@ -64,6 +69,11 @@ public partial class SettingsViewModel : PageViewModelBase
         "Dark",
         "System",
     };
+
+    [ObservableProperty] private CultureInfo selectedLanguage;
+
+    // ReSharper disable once MemberCanBeMadeStatic.Global
+    public IReadOnlyList<CultureInfo> AvailableLanguages => Cultures.SupportedCultures;
 
     public IReadOnlyList<float> AnimationScaleOptions { get; } = new[]
     {
@@ -103,17 +113,20 @@ public partial class SettingsViewModel : PageViewModelBase
         IPrerequisiteHelper prerequisiteHelper,
         IPyRunner pyRunner,
         ServiceManager<ViewModelBase> dialogFactory,
-        SharedState sharedState)
+        SharedState sharedState,
+        ITrackedDownloadService trackedDownloadService)
     {
         this.notificationService = notificationService;
         this.settingsManager = settingsManager;
         this.prerequisiteHelper = prerequisiteHelper;
         this.pyRunner = pyRunner;
         this.dialogFactory = dialogFactory;
+        this.trackedDownloadService = trackedDownloadService;
         
         SharedState = sharedState;
         
         SelectedTheme = settingsManager.Settings.Theme ?? AvailableThemes[1];
+        SelectedLanguage = Cultures.GetSupportedCultureOrDefault(settingsManager.Settings.Language);
         RemoveSymlinksOnShutdown = settingsManager.Settings.RemoveFolderLinksOnShutdown;
         SelectedAnimationScale = settingsManager.Settings.AnimationScale;
         
@@ -141,6 +154,43 @@ public partial class SettingsViewModel : PageViewModelBase
             "Light" => ThemeVariant.Light,
             _ => ThemeVariant.Default
         };
+    }
+
+    partial void OnSelectedLanguageChanged(CultureInfo? oldValue, CultureInfo newValue)
+    {
+        if (oldValue is null || newValue.Name == Cultures.Current.Name) return;
+        // Set locale
+        if (AvailableLanguages.Contains(newValue))
+        {
+            Logger.Info("Changing language from {Old} to {New}", 
+                oldValue, newValue);
+
+            Cultures.TrySetSupportedCulture(newValue);
+            settingsManager.Transaction(s => s.Language = newValue.Name);
+            
+            var dialog = new BetterContentDialog
+            {
+                Title = Resources.Label_RelaunchRequired,
+                Content = Resources.Text_RelaunchRequiredToApplyLanguage,
+                DefaultButton = ContentDialogButton.Primary,
+                PrimaryButtonText = Resources.Action_Relaunch,
+                CloseButtonText = Resources.Action_RelaunchLater
+            };
+
+            Dispatcher.UIThread.InvokeAsync(async () =>
+            {
+                if (await dialog.ShowAsync() == ContentDialogResult.Primary)
+                {
+                    Process.Start(Compat.AppCurrentPath);
+                    App.Shutdown();
+                }
+            });
+        }
+        else
+        {
+            Logger.Info("Requested invalid language change from {Old} to {New}", 
+                oldValue, newValue);
+        }
     }
     
     partial void OnRemoveSymlinksOnShutdownChanged(bool value)
@@ -329,6 +379,50 @@ public partial class SettingsViewModel : PageViewModelBase
             "Stability Matrix has been added to the Start Menu for all users.", NotificationType.Success);
     }
 
+    public async Task PickNewDataDirectory()
+    {
+        var viewModel = dialogFactory.Get<SelectDataDirectoryViewModel>();
+        var dialog = new BetterContentDialog
+        {
+            IsPrimaryButtonEnabled = false,
+            IsSecondaryButtonEnabled = false,
+            IsFooterVisible = false,
+            Content = new SelectDataDirectoryDialog
+            {
+                DataContext = viewModel
+            }
+        };
+
+        var result = await dialog.ShowAsync();
+        if (result == ContentDialogResult.Primary)
+        {
+            // 1. For portable mode, call settings.SetPortableMode()
+            if (viewModel.IsPortableMode)
+            {
+                settingsManager.SetPortableMode();
+            }
+            // 2. For custom path, call settings.SetLibraryPath(path)
+            else
+            {
+                settingsManager.SetLibraryPath(viewModel.DataDirectory);
+            }
+            
+            // Restart
+            var restartDialog = new BetterContentDialog
+            {
+                Title = "Restart required",
+                Content = "Stability Matrix must be restarted for the changes to take effect.",
+                PrimaryButtonText = "Restart",
+                DefaultButton = ContentDialogButton.Primary,
+                IsSecondaryButtonEnabled = false,
+            };
+            await restartDialog.ShowAsync();
+            
+            Process.Start(Compat.AppCurrentPath);
+            App.Shutdown();
+        }
+    }
+
     #endregion
     
     #region Debug Section
@@ -405,6 +499,32 @@ public partial class SettingsViewModel : PageViewModelBase
     {
         // Use try-catch to generate traceback information
         throw new OperationCanceledException("Example Message");
+    }
+
+    [RelayCommand]
+    private async Task DebugTrackedDownload()
+    {
+        var textFields = new TextBoxField[]
+        {
+            new()
+            {
+                Label = "Url",
+            },
+            new()
+            {
+                Label = "File path"
+            }
+        };
+
+        var dialog = DialogHelper.CreateTextEntryDialog("Add download", "", textFields);
+
+        if (await dialog.ShowAsync() == ContentDialogResult.Primary)
+        {
+            var url = textFields[0].Text;
+            var filePath = textFields[1].Text;
+            var download = trackedDownloadService.NewDownload(new Uri(url), new FilePath(filePath));
+            download.Start();
+        }
     }
     #endregion